Remove None/noToken.
Moving this into a publicly importable package makes me worried about
clients of the package accidentally leaking noToken; the package can't
actually prevent clients from leaking it.

Instead, we just generate a new token and don't share it in the places
where we were using noToken. This is almost the same, except that a
couple things need to move to account for the fact that New() can return
an error.